LayoutTests/fast/forms/focus-option-control-on-page-expected.txt

https://bugs.webkit.org/show_bug.cgi?id=68412 - This test checks to see if option(alt)-tabbing properly focuses form elements that are normally not focused. For testing, the assumption is that by default pressing tab will skip over buttons, and option-tab will include buttons.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Pressing tab 4 times:
PASS result is " /1:focused text field /2: /3:focused text field /4:"
Pressing shift-tab 4 times:
PASS result is " /1:focused text field /2: /3:focused text field /4:"
Pressing option-tab 4 times:
PASS result is " /1:focused first button /2:focused text field /3:focused second button /4:"
Pressing shift-option-tab 4 times:
PASS result is " /1:focused second button /2:focused text field /3:focused first button /4:"
